[IMP][V2.2] mrp: improve tests
mrp tests were only an dumb override of the base tests

-use monkey patch for override
-split into 2 files, to split tests specific to mrp
