+++ electricity_mrp/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_mrp_electricity
+from . import test_electricity_contract
